streamer: refine code, add more test cases for fileSizeUpdated, relaySubDirUpdated
