+++ pkg/streamer/util_test.go
@@ -14,18 +14,9 @@
 package streamer
 
 import (
-	"context"
-	"io/ioutil"
-	"os"
-	"path"
-	"path/filepath"
-	"sync"
 	"testing"
-	"time"
 
 	. "github.com/pingcap/check"
-
-	"github.com/pingcap/dm/pkg/utils"
 )
 
 var _ = Suite(&testUtilSuite{})
@@ -35,149 +26,6 @@ func TestSuite(t *testing.T) {
 }
 
 type testUtilSuite struct {
-}
-
-func (t *testUtilSuite) TestRelaySubDirUpdated(c *C) {
-	var (
-		relayFiles = []string{
-			"mysql-bin.000001",
-			"mysql-bin.000002",
-			"mysql-bin.000003",
-			"mysql-bin.000004",
-		}
-		relayPaths      = make([]string, len(relayFiles))
-		data            = []byte("meaningless file content")
-		size            = int64(len(data))
-		watcherInterval = 100 * time.Millisecond
-	)
-
-	// create relay log dir
-	subDir, err := ioutil.TempDir("", "test_relay_sub_dir_updated")
-	c.Assert(err, IsNil)
-	defer os.RemoveAll(subDir)
-
-	// join the path
-	for i, rf := range relayFiles {
-		relayPaths[i] = path.Join(subDir, rf)
-	}
-
-	// create first file
-	f, err := os.Create(relayPaths[0])
-	c.Assert(err, IsNil)
-	f.Close()
-
-	// 1. watcher, for file updated
-	var wg sync.WaitGroup
-	wg.Add(1)
-	go func() {
-		defer wg.Done()
-		latestFileSize := int64(0)
-		up, err2 := relaySubDirUpdated(context.Background(), watcherInterval, subDir, relayPaths[0], relayFiles[0], latestFileSize)
-		c.Assert(err2, IsNil)
-		c.Assert(up, Equals, relayPaths[0])
-	}()
-
-	// update file
-	f, err = os.OpenFile(relayPaths[0], os.O_WRONLY|os.O_APPEND, 0666)
-	c.Assert(err, IsNil)
-	f.Write(data)
-	f.Close()
-	wg.Wait()
-
-	// 2. create and update file
-	f, err = os.Create(relayPaths[1])
-	c.Assert(err, IsNil)
-	f.Write(data)
-	f.Close()
-
-	// watcher, but file size has changed
-	wg.Add(1)
-	go func() {
-		defer wg.Done()
-		latestFileSize := size - 1 // less than file size
-		up, err2 := relaySubDirUpdated(context.Background(), watcherInterval, subDir, relayPaths[1], relayFiles[1], latestFileSize)
-		c.Assert(err2, IsNil)
-		c.Assert(up, Equals, relayPaths[1])
-	}()
-	wg.Wait()
-
-	// 3. watcher, for file created
-	wg.Add(1)
-	go func() {
-		defer wg.Done()
-		latstFileSize := size
-		up, err2 := relaySubDirUpdated(context.Background(), watcherInterval, subDir, relayPaths[1], relayFiles[1], latstFileSize)
-		c.Assert(err2, IsNil)
-		c.Assert(up, Equals, relayPaths[2])
-	}()
-
-	// create file
-	f, err = os.Create(relayPaths[2])
-	c.Assert(err, IsNil)
-	f.Close()
-	wg.Wait()
-
-	// 4. create file
-	f, err = os.Create(relayPaths[3])
-	c.Assert(err, IsNil)
-	f.Close()
-
-	// watcher, for newer file already generated
-	wg.Add(1)
-	go func() {
-		defer wg.Done()
-		latstFileSize := int64(0)
-		up, err := relaySubDirUpdated(context.Background(), watcherInterval, subDir, relayPaths[2], relayFiles[2], latstFileSize)
-		c.Assert(err, IsNil)
-		c.Assert(up, Equals, relayPaths[3])
-	}()
-	wg.Wait()
-}
-
-func (t *testUtilSuite) TestFileSizeUpdatedError(c *C) {
-	var (
-		relayFile       = "mysql-bin.000001"
-		data            = []byte("meaningless file content")
-		watcherInterval = 100 * time.Millisecond
-	)
-
-	// create relay log dir
-	subDir, err := ioutil.TempDir("", "test_file_size_updated_error")
-	c.Assert(err, IsNil)
-	defer os.RemoveAll(subDir)
-
-	// create relay log
-	relayPath := path.Join(subDir, relayFile)
-	err = ioutil.WriteFile(relayPath, data, 0644)
-	c.Assert(err, IsNil)
-
-	// watcher, for file size decrease
-	var wg sync.WaitGroup
-	wg.Add(1)
-	go func() {
-		defer wg.Done()
-		latestFileSize := int64(len(data))
-		_, err2 := relaySubDirUpdated(context.Background(), watcherInterval, subDir, relayPath, relayFile, latestFileSize)
-		c.Assert(err2, NotNil)
-	}()
-
-	// truncate relay log
-	err = ioutil.WriteFile(relayPath, []byte(""), 0644)
-	c.Assert(err, IsNil)
-	wg.Wait()
-
-	// watcher, for file not found
-	wg.Add(1)
-	go func() {
-		defer wg.Done()
-		latestFileSize := int64(0)
-		_, err2 := relaySubDirUpdated(context.Background(), watcherInterval, subDir, relayPath, relayFile, latestFileSize)
-		c.Assert(err2, NotNil)
-	}()
-
-	err = os.Remove(relayPath)
-	c.Assert(err, IsNil)
-	wg.Wait()
 }
 
 func (t *testUtilSuite) TestGetNextUUID(c *C) {
@@ -240,56 +88,4 @@ func (t *testUtilSuite) TestGetNextUUID(c *C) {
 		c.Assert(nu, Equals, cs.nextUUID)
 		c.Assert(nus, Equals, cs.nextUUIDSuffix)
 	}
-}
-
-func (t *testUtilSuite) TestGetFirstBinlogName(c *C) {
-	var (
-		baseDir = c.MkDir()
-		uuid    = "b60868af-5a6f-11e9-9ea3-0242ac160006.000001"
-		subDir  = filepath.Join(baseDir, uuid)
-	)
-
-	// sub directory not exist
-	name, err := getFirstBinlogName(baseDir, uuid)
-	c.Assert(err, ErrorMatches, ".*no such file or directory.*")
-	c.Assert(name, Equals, "")
-
-	// empty directory
-	err = os.MkdirAll(subDir, 0744)
-	c.Assert(err, IsNil)
-	name, err = getFirstBinlogName(baseDir, uuid)
-	c.Assert(err, ErrorMatches, ".*not found.*")
-	c.Assert(name, Equals, "")
-
-	// has file, but not valid binlog file
-	filename := "invalid.bin"
-	err = ioutil.WriteFile(filepath.Join(subDir, filename), nil, 0644)
-	c.Assert(err, IsNil)
-	name, err = getFirstBinlogName(baseDir, uuid)
-	c.Assert(err, ErrorMatches, ".*not valid.*")
-	err = os.Remove(filepath.Join(subDir, filename))
-	c.Assert(err, IsNil)
-
-	// has a valid binlog file
-	filename = "z-mysql-bin.000002" // z prefix, make it become not the _first_ if possible.
-	err = ioutil.WriteFile(filepath.Join(subDir, filename), nil, 0644)
-	c.Assert(err, IsNil)
-	name, err = getFirstBinlogName(baseDir, uuid)
-	c.Assert(err, IsNil)
-	c.Assert(name, Equals, filename)
-
-	// has one more earlier binlog file
-	filename = "z-mysql-bin.000001"
-	err = ioutil.WriteFile(filepath.Join(subDir, filename), nil, 0644)
-	c.Assert(err, IsNil)
-	name, err = getFirstBinlogName(baseDir, uuid)
-	c.Assert(err, IsNil)
-	c.Assert(name, Equals, filename)
-
-	// has a meta file
-	err = ioutil.WriteFile(filepath.Join(subDir, utils.MetaFilename), nil, 0644)
-	c.Assert(err, IsNil)
-	name, err = getFirstBinlogName(baseDir, uuid)
-	c.Assert(err, IsNil)
-	c.Assert(name, Equals, filename)
 }
